ContainsNearbyAlmostDuplicate

+++ Java/src/main/java/LeetCode_2019_06/LeetCode20190620/ContainsNearbyAlmostDuplicate.java
@@ -38,7 +38,7 @@ public class ContainsNearbyAlmostDuplicate {
         }
         for(int i = 0; i < nums.length;i++){
             for(int j = i + 1; j < nums.length && j <= i + k;j++){
-                if(Math.abs((long)(nums[i] - nums[j])) <= t){
+                if(Math.abs((long)nums[i] - (long)nums[j]) <= t){
                     return true;
                 }
             }
